feat: added custom hook: useStatusEvents
